StudentAdditionalTrainingListController moved and renamed 	renamed:    app/Http/Controllers/api/StudentAdditionalTrainingListController, -> app/Http/Controllers/api/Student/StudentAdditionalTrainingListController.php 	renamed:    app/Service/StudentAdditionalTrainingService.php -> app/Service/StudentAdditionalTrainingListService.php 	modified:   routes/api/v1.php 	modified:   tests/Feature/Service/StudentAdditionalTrainingListServiceTest.php 	modified:   tests/Feature/Student/StudentAdditionalTrainingListControllerTest.php
